Overview

RandomizedDiagonalEstimation.jl is a Julia package that implements randomized methods for the estimation of the diagonal of matrices of matrices and matrix functions. For pure matrix diagonal estimation we provide the following algorithms

Girard-Hutchinson Estimator [1]

Diag++ [2]

NysDiag++

XDiag [3]

Full Hutchinson Shifts

For the estimation of the diagonal of matrix function we combine the Girard-Hutchinson Estimator with the following approximations for $f(\textbf{A})\textbf{b}$

Chebyshev interpolants to approximate $f$ on the interval $[\lambda_{min},\lambda_{max}]$

Minimax polynomials from the Remez algorithm to approximate $f$ on the interval $[\lambda_{min},\lambda_{max}]$

Arnoldi approximations

The package provides three functions: EstimateDiagonal, EstimateFunctionDiagonal and EstimateMoMDiagonal. The last function incorporates the median of means package into diagonal estimation. A more detailed elaboration of the algorithms and theoretical properties can be found in this thesis: (will be updated once available on DIVA)

Estimating the Diagonal of Matrix

The estimation of the diagonal of a function of a square matrix is handled by the function EstimateDiagonal

RandomizedDiagonalEstimation.EstimateDiagonal — Function

function EstimateDiagonal(A::Matrix{Float64},Algorithm::Symbol, StoppingCriterion::Symbol, distribution::Symbol, normalizationParam::Bool=true, parallelizationParam::Bool=false ;maxqueries::Int=0,O=nothing,var_est::Float64=1/eps(),epsilon::Float64=1.0, delta::Float64=1.0,con::Float64=1.0)

Main function to compute a randomized estimate of the diagonal of a matrix A

Input:

A: Matrix to estimate the diagonal of

Algorithm: Choose which diagonal estimator is used, options are

:GirardHutchinson

DiagPP

:NysDiagPP

:XDiag

:GirardHutchinsonShift

:RSVD

:AdaptiveDiagPP

StoppingCriterion: How to terminate, possible options

doubling: Use doubling strategy and terminate when the relative error estimate is below a threshold eps, required parameter

queries_start: number of queries to start with

eps: bound for the relative error estimate

Note: has only been implemented with :GirardHutchinson and :DiagPP, other methods will be added with future releases

queries: terminate when the maximum number of queries to A is reacher

maxqueries: maximum number of queries to A

adaptive: Using an epsilon delta estimator with :AdaptiveDiagPP

epsdelta: parameters (epsilon,delta) for an epsilon-delta estimator

maxiter: maximum iterations

distribution: Select the distribution from which the random vectors are drawn, inbuilt options are

:Rademacher

:Gaussion

:custom, in this case provide a matrix with the test vectors as columns

O: matrix with the test vectors as columns

Estimating the Diagonal of Matrices Using Median of Means

EstimateMoMDiagonal combines the functionality of EstimateDiagonal with the median of means principle.

RandomizedDiagonalEstimation.EstimateMoMDiagonal — Function

function EstimateMoMDiagonal(A::Matrix{Float64},Algorithm::Symbol, StoppingCriterion::Symbol, distribution::Symbol, ngroups::Int, groupsize::Int, normalizationParam::Bool=true, parallelizationParam::Bool=false ;maxqueries::Int,O=nothing)

Function that computes a randomized estimate of the diagonal of a matrix using the median of means principle

Input:

A: Matrix to estimate the diagonal of

Algorithm: Choose which diagonal estimator is used, options are

:GirardHutchinson

:DiagPP

:NysDiagPP

:XDiag

:XNysDiag

:GirardHutchinsonShift

:RSVD

:AdaptiveDiagPP

StoppingCriterion: How to terminate, possible options

queries: terminate when the maximum number of queries to A is reacher

maxqueries: maximum number of queries to A

Note potential further stopping criteria will be added in future versions

distribution: Select the distribution from which the random vectors are drawn, inbuilt options are

:Rademacher

:Gaussion

:custom, in this case provide a matrix with the test vectors as columns

O: matrix with the test vectors as columns

ngroups: Number of groups for the Median of Means estimator

groupsize: Groupsize for the Median of Means estimator

Estimating the Diagonal of Matrix

The estimation of the diagonal of a function of a square matrix is handled by the function EstimateDiagonal

RandomizedDiagonalEstimation.EstimateFunctionDiagonal — Function

function EstimateFunctionDiagonal(A::Matrix{Float64},f,Algorithm::Symbol, StoppingCriterion::Symbol, distribution::Symbol, MatFuncApprox::Symbol, deg::Int64, normalizationParam::Bool=true;maxqueries::Int,int::Tuple=(0.0,1.0),O=nothing)

Main function to compute a randomized estimate of the diagonal of a matrix function

Input:

A: Matrix as input for the function

f: Function to approximate diagonal of

Algorithm: Choose which diagonal estimator is used, options are

:GirardHutchinson

MatFuncApprox: How to approximate f(A)b

Chebshev: Using Chebyshev polynomials

requires interval int and degree deg

Remez: Using Remez polynomials

requires interval int and degree deg

Krylov: Using the Arnoldi approximation

requires maximum potency of A in the Krylov subspace denoted by deg

CG: Use conjugate gradient method to approximate diagonal of the inverse, Attention: f is neglected

requires maximum potency of A in the Krylov subspace denoted by deg

StoppingCriterion: How to terminate, possible options

doubling: Use doubling strategy and terminate when the relative error estimate is below a threshold eps, required parameter

queries_start: number of queries to start with

eps: bound for the relative error estimate

queries: terminate when the maximum number of queries to A is reacher

maxqueries: maximum number of queries to A

distribution: Select the distribution from which the random vectors are drawn, inbuilt options are

:Rademacher

:Gaussion

:custom, in this case provide a matrix with the test vectors as columns

O: matrix with the test vectors as columns

deg: degree of polynomial or maximum potency of A in the Krylov subspace, depending on MatFuncApprox

Citing RandomizedDiagonalEstimation.jl

If you use RandomizedDiagonalEstimation.jl for academic research and wish to cite it, please cite the following paper (will be updated).

References

[1] C. Bekas, E. Kokiopoulou, and Y. Saad. “An estimator for the diagonal of a matrix”. In: Applied Numerical Mathematics 57.11 (2007). Numerical Algorithms, Parallelism and Applications (2), pp. 1214–1229. issn: 0168-9274.
[2] R. A. Baston and Y. Nakatsukasa. “Stochastic diagonal estimation: probabilistic bounds and an improved algorithm”. In: ArXiv abs/2201.10684 (2022).
[3] E. N. Epperly and J. A. Tropp. Efficient error and variance estimation for randomized matrix com- putations. 2023. arXiv: 2207.06342 [math.NA].